pyproject.toml
[tool.poetry]
name = "dentiai-homework"
version = "0.1.0"
description = "Test task for Senior Software Engineer position in Denti.ai"
authors = ["Pavel Pavlov <ppavlov.swe@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.14.0"
django-treebeard = "^4.7"
psycopg2 = "^2.9.9"
pillow = "^10.1.0"
pytest-django = "^4.7.0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.18.1"
mypy = "^1.7.1"
flake8 = "^6.1.0"
black = "^23.12.0"
pytest = "^7.4.3"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "dentiai.settings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
